Extract Scala object which defines GraphQL schema
Split GraphQlServer object: keep the query execution code in GraphQlServer,
but create a new object to define the types and the schemas.

This will make it simpler to add a step which generates a schema text file.
